Title: 합과 차의 나눗셈, Time: 150ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240203/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -7,6 +7,6 @@ public class Main {
         int a = sc.nextInt();
         int b = sc.nextInt();
 
-        System.out.printf("%.2f", (float)(a + b) / (a - b));
+        System.out.printf("%.2f", (double)(a + b) / (a - b));
     }
 }
